1.工单附件下载 2.受理/解决/关闭/重开 待实现

+++ src/main/java/evdc/vianet/ticket/service/TicketCommentService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import evdc.vianet.ticket.entity.TicketMessageView;
 
 public interface TicketCommentService {
-	public int addTicketComment(long ticketId, long userId, long teamId, String message, String scope);
+	public long addTicketComment(long ticketId, long userId, long teamId, String message, String scope);
 	public List<TicketMessageView> getCommentsByTicketIdAndScope(long ticketId, String scope);
 	public List<TicketMessageView> getCommentsByTicketIdAndteamIdAndScope(long ticketId, long teamId, String scope);
 }
